Extracted manage journals abbreviations controller
