+++ src/main/resources/net/sf/jabref/gui/journals/ManageJournalAbbreviations.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
 
-<DialogPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="net.sf.jabref.gui.journals.ManageJournalAbbreviationsView">
+<DialogPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="net.sf.jabref.gui.journals.ManageJournalAbbreviationsController">
    <content>
       <BorderPane prefHeight="200.0" prefWidth="200.0">
          <center>
